requirements.txt
streamlit==1.32.0
langchain-community==0.0.27
langchain-core==0.1.27
faiss-cpu==1.7.4
sentence-transformers==2.5.1
google-generativeai==0.3.2 